Wire up scheduling child entries
